feat #22 - 주차별 지출 store 생성 및 analysisStore로 파일 이름 변경
